Move everything to separate directory to not clog App Router, #12
